src/requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-passwordreset==1.0.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
drf-yasg==1.16.0
gunicorn==19.4.5
oauth2-provider==0.0
oauthlib==3.0.1
python-memcached==1.59
requests==2.22.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
stripe==2.21.0

